refactor(components): rename files :recycle:
